fw-4255, move permissions manager
